pyproject.toml
[tool.poetry]
name = "rest-against-humanity"
version = "2.2.0"
description = "A public API for Cards Against Humanity"
authors = ["celsius narhwal <celsiusnarhwal@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.92.0"
pydantic = "^1.10.7"
ujson = "^5.7.0"
strawberry-graphql = { extras = ["fastapi"], version = "0.176.3" }

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.19.0"
pyperclip = "^1.8.2"
rich-tracebacks = "^1.2.1"
uvicorn = "^0.20.0"

[tool.poetry.group.docs.dependencies]
cairosvg = "^2.7.0"
mike = "^1.1.2"
pillow = "^9.4.0"
poethepoet = "^0.19.0"
mkdocs-material = { git = "https://github.com/squidfunk/mkdocs-material-insiders" }

[tool.poetry.group.ci.dependencies]
semver = "^2.13.0"
tomlkit = "^0.11.6"

[tool.poetry.group.repl.dependencies]
ipython = "^8.11.0"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"

[tool.poe.tasks]
clear = "rm -rf site .cache"
docs = { shell = "mkdocs serve" }
export = "poetry export -f requirements.txt -o api/requirements.txt"
format = { shell = "for hook in black prettier isort; do pre-commit run $hook --all-files; done" }

[tool.poe.tasks.fontawesome]
shell = """
npm install --no-save @celsiusnarhwal/fontawesome
mv node_modules/@celsiusnarhwal/fontawesome/svgs docs/.overrides/.icons/fontawesome
rm -rf node_modules
"""

[tool.poe.tasks.build]
shell = """
mkdocs build
cp -r api vercel.json site
poetry export -f requirements.txt -o site/api/requirements.txt
"""

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
